+++ lib/src/badge_position.dart
@@ -16,8 +16,23 @@ class BadgePosition {
   /// This is the distance to the bottom side of a parent widget.
   final double? bottom;
 
+  /// This is indicate that centralise is required
+  final bool isCenter;
+
   /// This creates a new instance of this widget
-  const BadgePosition({this.top, this.end, this.bottom, this.start});
+  const BadgePosition({
+    this.top,
+    this.end,
+    this.bottom,
+    this.start,
+    this.isCenter = false,
+  });
+
+  /// Factory method that creates a new instance of this widget
+  /// according to the center
+  factory BadgePosition.center() {
+    return BadgePosition(isCenter: true);
+  }
 
   /// Factory method that creates a new instance of this widget
   /// according to [top] and [start]